+++ fv3net/regression/__init__.py
@@ -1,4 +1,3 @@
 from . import loaders
-from .constants import TIME_NAME, SAMPLE_DIM_NAME
 
-__all__ = ["loaders", "TIME_NAME", "SAMPLE_DIM_NAME"]
+__all__ = ["loaders"]