+++ fv3net/regression/constants.py
@@ -1,6 +1,0 @@
-TIME_NAME = "time"
-SAMPLE_DIM_NAME = "sample"
-Z_DIM_NAME = "z"
-
-
-__all__ = ["TIME_NAME", "SAMPLE_DIM_NAME", "Z_DIM_NAME"]