+++ fv3net/regression/sklearn/train.py
@@ -3,10 +3,12 @@ import joblib
 import logging
 import os
 import yaml
+import xarray as xr
 from dataclasses import dataclass
 from typing import Iterable, Sequence
 
 from .. import loaders
+from ..loaders._batch import mapper_to_batches
 from .wrapper import SklearnWrapper, RegressorEnsemble
 from sklearn.compose import TransformedTargetRegressor
 from sklearn.preprocessing import StandardScaler
@@ -25,7 +27,7 @@ class ModelTrainingConfig:
     hyperparameters: dict
     input_variables: Iterable[str]
     output_variables: Iterable[str]
-    batch_function: str
+    mapping_function: str
     batch_kwargs: dict
 
 
@@ -46,18 +48,21 @@ def load_model_training_config(config_path: str) -> ModelTrainingConfig:
     return ModelTrainingConfig(**config_dict)
 
 
-def load_data_sequence(data_path: str, train_config: ModelTrainingConfig) -> Sequence:
+def load_data_sequence(
+    data_path: str, train_config: ModelTrainingConfig
+) -> Sequence[xr.Dataset]:
     """
     Args:
         data_path: data location
         train_config: model training configuration
 
     Returns:
-        sequence: xr datasets for training batches
+        Sequence of xarray datasets
     """
-    batch_function = getattr(loaders, train_config.batch_function)
-    ds_batches = batch_function(
-        data_path,
+    mapping_function = getattr(loaders, train_config.mapping_function)
+    data_mapping = mapping_function(data_path)
+    ds_batches = mapper_to_batches(
+        data_mapping,
         list(train_config.input_variables) + list(train_config.output_variables),
         **train_config.batch_kwargs,
     )
@@ -102,7 +107,7 @@ def _get_transformed_batch_regressor(train_config):
     return model_wrapper
 
 
-def train_model(batched_data: Sequence, train_config: ModelTrainingConfig):
+def train_model(batched_data: Sequence[xr.Dataset], train_config: ModelTrainingConfig):
     """
     Args:
         batched_data: training batch datasets